+++ ag41_project_sample1/Data.h
@@ -12,28 +12,26 @@ using namespace std;
 #include <map>
 #include "Client.h"
 #include "Action.h"
-namespace Donnees
-{
-	class Data {
-		map<pair<Client*, Client*>, int> distanceClients;
-		map<Client*, int> distanceFournisseur;
-		vector<Client*> listeClient;
-		map<int, vector<Action*>* > sol;
-		float kTransport;
-	public:
-		Data();
-		virtual ~Data();
-		Data(Data &d);
-		vector<Client*> getListeClient();
-		void ajouterClient(Client* c);
-		int distanceClient(Client* c);
-		void distanceClient(Client* c, int dist);
-		int distanceClient(Client* c1, Client* c2);
-		void distanceClient(Client* c1, Client* c2, int dist);
-		float getKTransport();
-		void setKTransport(float k);
+namespace Donnees {
+class Data {
+	map<pair<Client*, Client*>, double> distanceClients;
+	map<Client*, double> distanceFournisseur;
+	vector<Client*> listeClient;
+	map<double, vector<Action*>*> sol;
+	double kTransport;
+public:
+	Data();
+	virtual ~Data();
+	Data(Data &d);
+	vector<Client*> getListeClient();
+	void ajouterClient(Client* c);
+	double distanceClient(Client* c);
+	void distanceClient(Client* c, double dist);
+	double distanceClient(Client* c1, Client* c2);
+	void distanceClient(Client* c1, Client* c2, double dist);
+	double getKTransport();
+	void setKTransport(double k);
 
-
-	};
+};
 }
 #endif /* DATA_H_ */
